IOS-4070: Split params and result DTOs into separate files, update some DTOs
