+++ BlockchainSdk/WalletManagers/NEAR/DTO/Params/NEARNetworkParams.swift
@@ -1,0 +1,12 @@
+//
+//  NEARNetworkParams.swift
+//  BlockchainSdk
+//
+//  Created by Andrey Fedorov on 13.10.2023.
+//  Copyright © 2023 Tangem AG. All rights reserved.
+//
+
+import Foundation
+
+/// A top-level namespace.
+enum NEARNetworkParams {}
