+++ BlockchainSdk/WalletManagers/NEAR/DTO/Result/NEARNetworkResult.TransactionSendAsync.swift
@@ -1,0 +1,14 @@
+//
+//  NEARNetworkResult.TransactionSendAsync.swift
+//  BlockchainSdk
+//
+//  Created by Andrey Fedorov on 26.10.2023.
+//  Copyright © 2023 Tangem AG. All rights reserved.
+//
+
+import Foundation
+
+extension NEARNetworkResult {
+    /// Hash of the transaction.
+    typealias TransactionSendAsync = String
+}
